# docs/swagger.yaml
basePath: /
definitions:
  main.HealthResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  main.WelcomeResponse:
    properties:
      description:
        type: string
      docs:
        type: string
      endpoints:
        items:
          type: string
        type: array
      name:
        type: string
      version:
        type: string
    type: object
  models.BestSeller:
    properties:
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: integer
    type: object
  models.Category:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.CategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  models.CheckoutItem:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  models.CheckoutRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/models.CheckoutItem'
        type: array
    type: object
  models.Product:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
    type: object
  models.ProductRequest:
    properties:
      category_id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
    type: object
  models.SalesReport:
    properties:
      best_seller:
        $ref: '#/definitions/models.BestSeller'
      total_revenue:
        type: integer
      total_transactions:
        type: integer
    type: object
  models.Transaction:
    properties:
      created_at:
        type: string
      details:
        items:
          $ref: '#/definitions/models.TransactionDetail'
        type: array
      id:
        type: integer
      total_amount:
        type: integer
    type: object
  models.TransactionDetail:
    properties:
      id:
        type: integer
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: integer
      subtotal:
        type: integer
      transaction_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for Point of Sale (POS) system with Layered Architecture
  title: Kasir API
  version: "2.0"
paths:
  /:
    get:
      description: Display welcome information and endpoint list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.WelcomeResponse'
      summary: Welcome
      tags:
      - Info
  /categories:
    get:
      description: Mengambil semua daftar kategori
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Membuat kategori baru
      parameters:
      - description: Data kategori
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Category'
      summary: Create category
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Menghapus kategori berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete category
      tags:
      - Categories
    get:
      description: Mengambil kategori berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "404":
          description: Category tidak ditemukan
          schema:
            type: string
      summary: Get category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Mengupdate kategori berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data kategori
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      summary: Update category
      tags:
      - Categories
  /health:
    get:
      description: Check if the API is running properly
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.HealthResponse'
      summary: Health check
      tags:
      - Health
  /products:
    get:
      description: Get all products from the database, optionally filtered by name
      parameters:
      - description: Search by product name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
      summary: Create product
      tags:
      - Products
  /products/{id}:
    delete:
      description: Delete a product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete product
      tags:
      - Products
    get:
      description: Get a product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "404":
          description: Product not found
          schema:
            type: string
      summary: Get product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update a product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Update product
      tags:
      - Products
  /reports:
    get:
      description: Get sales report filtered by start_date and end_date (YYYY-MM-DD)
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SalesReport'
      summary: Get sales report with custom date range
      tags:
      - Reports
  /reports/today:
    get:
      description: Get total revenue, total transactions, and best seller for today
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SalesReport'
      summary: Get sales report for today
      tags:
      - Reports
  /transactions:
    post:
      consumes:
      - application/json
      description: Create a new transaction with multiple items
      parameters:
      - description: Checkout data
        in: body
        name: checkout
        required: true
        schema:
          $ref: '#/definitions/models.CheckoutRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Transaction'
      summary: Create transaction (checkout)
      tags:
      - Transactions
swagger: "2.0"